+++ applications/vessel-history/src/routes/routes.ts
@@ -17,6 +17,7 @@ import { UpdateQueryParamsAction } from './routes.actions'
 export const HOME = 'HOME'
 export const LOGIN = 'LOGIN'
 export const PROFILE = 'PROFILE'
+export const SETTINGS = 'SETTINGS'
 
 export type ROUTE_TYPES = typeof HOME | typeof PROFILE
 
@@ -28,6 +29,10 @@ const thunk = async (
 const routesMap: RoutesMap = {
   [HOME]: {
     path: '/',
+    thunk,
+  },
+  [SETTINGS]: {
+    path: '/settings',
     thunk,
   },
   [LOGIN]: {