+++ packages/dataviews-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @globalfishingwatch/dataviews-client
 
+## 8.0.0
+
+### Patch Changes
+
+- Updated dependencies [6c1b25bb]
+  - @globalfishingwatch/layer-composer@7.1.0
+
 ## 7.0.0
 
 ### Major Changes